requirements.txt
streamlit>=1.20.0
langchain>=0.0.200
langchain-community>=0.1.0
faiss-cpu>=1.7.3
openai>=0.27.0
python-dotenv>=1.0.0
llama-cpp-python>=0.1.80   # only if you plan to use a local LLaMA
boto3>=1.26.0              # for AWS S3 access
